==> main_analyses/15_biodiversity_uncertainty.Rmd <==
---
title: "Biodiversity Uncertainty"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
---

```{r}
library(raster)
library(sf)
library(tidyverse)
library(patchwork)
library(parallel)
library(foreach)
library(doParallel)
library(furrr)

source(here::here("common.R"))

variant_name <- "15_biodiversity_uncertainty"

variant_dir <- file.path(emLab_project_dir,"data", "03_output", variant_name)

dir.create(variant_dir)
```

# Load data 

```{r}
features_df <- data.table::fread(file.path(emLab_project_dir, "data", "02_processed", "features_df", "all_bio_features_df.csv"))

feature_names <- colnames(features_df)[c(-1,-2,-3)]

n_features <- length(feature_names)
```

```{r mpas}
MPAs <- raster(file.path(emLab_project_dir, "data", "02_processed", "masks", "fully_highly_protected_reviewed_MPAs.tif"))

MPAs_df <- MPAs %>% 
  raster::as.data.frame(xy = T) %>%
  inner_join(ocean_matrix) %>% 
  select(x, y, cell_id, mpa = fully_highly_protected_reviewed_MPAs) %>%
  as_tibble() 
```

```{r impacts}
bio_abatable_impacts_df <- raster(file.path(emLab_project_dir, 
                                            "data", "02_processed", "impacts", "chi", "abatable_impacts_5_yr_avg_log.tif")) %>% 
  raster::as.data.frame(xy = T) %>% 
  set_names(c("x", "y", "Ia")) %>% 
  inner_join(ocean_matrix) %>% 
  as_tibble() %>% 
  replace_na(list(Ia = 0))

bio_unabatable_impacts_df <- raster(file.path(emLab_project_dir, 
                                              "data", "02_processed", "impacts", "chi", "unabatable_impacts_5_yr_avg_log.tif")) %>% 
  raster::as.data.frame(xy = T) %>% 
  set_names(c("x", "y", "Iu")) %>% 
  inner_join(ocean_matrix) %>% 
  as_tibble()%>% 
  replace_na(list(Iu = 0))
```


## Feature weights 

```{r}
spp_wts <- data.table::fread(file.path(emLab_project_dir, "data", "02_processed", "species_list", "spp_weights.csv")) %>% 
  as_tibble() %>% 
  rename(w = spp_weight)

smts_wts <- tibble(filepath = list.files(file.path(emLab_project_dir, "data", "02_processed","seamounts"), full.names = T)) %>% 
  mutate(w = sum(spp_wts$w)/n())

provs_wts <- tibble(filepath = list.files(file.path(emLab_project_dir, "data", "02_processed", "biogeography"), full.names = T, pattern = "\\.tif")) %>% 
  mutate(w = sum(spp_wts$w)/n())

feature_wts <- c(spp_wts$w, smts_wts$w, provs_wts$w)

names(feature_wts) <- feature_names

sum(spp_wts$w) == sum(smts_wts$w)

sum(smts_wts$w) == sum(provs_wts$w)
```

```{r}
features_matrix <- features_df %>% 
  select(-x,-y,-cell_id) %>% 
  as.matrix()

rownames(features_matrix) <- features_df$cell_id
```

## Algorithm

```{r}
random_cell_removal <- function(v, f){
  
  n_effective_cells <- sum(v > 0, na.rm = T)
  
  effective_cells <- which(v > 0)
  
  indices_to_remove <- sample(effective_cells, round(n_effective_cells*f))
  
  v[indices_to_remove] <- NA
  
  v
}
```

Do what follows 1000 times

```{r}
gc()

registerDoParallel(cores = 15)

foreach (i = c(500:1000)) %dopar% {

# RANDOMIZE

feature_wts <- feature_wts + runif(n = length(feature_wts), min = -sd(feature_wts), max = sd(feature_wts))

feature_wts[feature_wts < 0] <- 0

removed_fraction <-  runif(n = 1, min = 0, max = 0.3)

features_matrix <- features_matrix %>% 
  apply(MARGIN = 2, FUN = random_cell_removal, f = removed_fraction)

# stopifnot(
#   sum(map_lgl(features_df, is.numeric)) == ncol(features_df) 
#   ) # ALL numeric?

features_matrix <- sweep(features_matrix, 2, colSums(features_matrix, na.rm = T), FUN = "/")

# stopifnot(
#   sum(colSums(features_matrix, na.rm = T)) == ncol(features_matrix)
#   )  # Normalized OK?

features_matrix <- features_matrix[rowSums(is.na(features_matrix)) != ncol(features_matrix), , drop = F]

rownames(features_matrix) <- features_df$cell_id

# stopifnot(
#   identical(colnames(features_matrix), 
#           features_df %>% 
#             select(-x,-y,-cell_id) %>% 
#             colnames())
#   )  # Is the order of the features mantained?

features_matrix[is.na(features_matrix)] <- 0

# Get MPA vector

MPA_vect <- rownames(features_matrix) %>% 
  enframe() %>% 
  select(cell_id = value) %>% 
  left_join(MPAs_df) %>% 
  mutate(mpa =  !is.na(mpa)) %>% pull(mpa)

# Apply impacts

remains_BAU <- features_matrix %>% 
  sweep(1, (1 - bio_abatable_impacts_df$Ia), FUN = "*") %>% 
  sweep(1, (1 - bio_unabatable_impacts_df$Iu), FUN = "*") %>% 
  colSums(na.rm = T)

remains_MPA <- features_matrix %>% 
  sweep(1, (1 - bio_unabatable_impacts_df$Iu), FUN = "*") %>% 
  colSums(na.rm = T)

features_matrix <- features_matrix %>% 
  sweep(1, (bio_abatable_impacts_df$Ia - bio_abatable_impacts_df$Ia*bio_unabatable_impacts_df$Iu), FUN = "*")

# Initial conditions

z <- 0.25

protected_cells <- matrix(MPA_vect, nrow = 1, ncol = nrow(features_matrix))

colnames(protected_cells) <- rownames(features_matrix)

baseline_state <- remains_BAU + protected_cells%*%features_matrix

max_slopes <- z*baseline_state^(z - 1)

features_matrix <- features_matrix[!protected_cells, ,drop = F]

# Algorithm 

ranking <- list()

current_pick <- 0

current_state <- baseline_state

while(nrow(features_matrix) > 0){
  
  current_pick <- sum(current_pick, 1)
  
  slopes <- feature_wts*pmin(max_slopes, z*current_state^(z - 1))
  
  delta <- features_matrix%*%t(as.matrix(slopes))
  
  best_cell_indeces <- doBy::which.maxn(delta, n = 100)
  
  best_cells <- delta[best_cell_indeces, ] %>% 
    enframe() %>% 
    set_names(c('cell_id', "delta")) %>% 
    mutate(pick_order = current_pick)
  
  ranking <- bind_rows(ranking, best_cells)
  
  current_state <- features_matrix[best_cell_indeces, , drop = F] %>% 
    colSums() +
    current_state
  
  features_matrix <- features_matrix[-best_cell_indeces, , drop = F]
  
}

write_rds(ranking, 
          here::here("uncertainty_runs", paste0(paste("run", i,"f", 100*round(removed_fraction, 2), sep = "_"), ".rds")))

}
```

# Results

```{r}
plan(multiprocess)

uncertainty_runs_files <- list.files(file.path(emLab_project_dir, "data", "03_output", "15_biodiversity_uncertainty", "uncertainty_runs"),
                                     full.names = T)

uncertainty_runs_results <- furrr::future_map_dfr(uncertainty_runs_files,
                                                  .f = function(x){
                                                    
                                                    xx <- read_rds(x)
                                                    
                                                    MPAs_df %>% 
                                                      filter(!is.na(mpa)) %>% 
                                                      select(cell_id) %>% 
                                                      bind_rows(xx) %>% 
                                                      mutate(is_mpa = is.na(delta)) %>% 
                                                      mutate(fraction_protected = 1/n(),
                                                             fraction_protected = cumsum(fraction_protected)) %>% 
                                                      select(cell_id, is_mpa, fraction_protected)
                                                  })


# Load original model

og_results <- read_rds(file.path(emLab_project_dir, "data", "03_output", "05_spp_wts_smts_provs_MPAs", "ranking_results.rds")) 

og_results <- MPAs_df %>% 
  filter(!is.na(mpa)) %>% 
  select(cell_id) %>% 
  bind_rows(og_results) %>% 
  mutate(fraction_protected = 1/n(),
         fraction_protected = cumsum(fraction_protected))
```

## Uncertainty in top 5%

```{r}
# Calculate fraction of runs in top 5%
og_top5_cells <- og_results %>% 
  filter(fraction_protected <= 0.05) %>% 
  pull(cell_id)

uncertainty_raster_top5 <- uncertainty_runs_results %>% 
  filter(cell_id %in% og_top5_cells) %>% 
  group_by(cell_id) %>% 
  summarise(fraction_in_top5 = sum(fraction_protected <= 0.05)/n()) %>% 
  left_join(ocean_matrix) %>% 
  select(x, y, fraction_in_top5) %>% 
  rasterFromXYZ(crs = crs(ocean_low_res_moll)) %>% 
  projectRaster(ocean_low_res_moll, method = 'ngb') %>% 
  mask(ocean_low_res_moll)

uncertainty_map_top5 <- tmap::tm_shape(uncertainty_raster_top5)+
  tmap::tm_raster(title = "",
                  palette  = viridis::viridis(n = 4, direction = 1),
                  breaks = c(0, 0.75, 0.9, 0.95, 1),
                  labels = c("<75", "75-90","90-95", ">95"),
                  legend.is.portrait = T, 
                  legend.reverse = T)+
  tmap::tm_shape(land_50_moll)+
  tmap::tm_fill(col = "grey", border.col = "transparent")+
  tmap::tm_shape(MPAs)+
  tmap::tm_credits("% of runs within top 5%")+
  tmap::tm_raster(col = "fully_highly_protected_reviewed_MPAs", palette = "lightblue", legend.show	= F)+
  tmap::tm_layout(title = "Biodiversity Conservation Priorities",
                  title.position = c("center", .95),
                  inner.margins = c(0.12, 0, 0.08, 0.04), 
                  frame = F,
                  legend.position = c(.99, "center"))

tmap::tmap_save(uncertainty_map_top5, 
                filename = file.path(variant_dir, "uncertainty_map_top5.png"),
                dpi = 300,  height = 4, width = 8 )
```


## Uncertainty in top 10%

```{r}
# Calculate fraction of runs in top 10%
og_top10_cells <- og_results %>% 
  filter(fraction_protected <= 0.1) %>% 
  pull(cell_id)

uncertainty_raster_top10 <- uncertainty_runs_results %>% 
  filter(cell_id %in% og_top10_cells) %>% 
  group_by(cell_id) %>% 
  summarise(fraction_in_top10 = sum(fraction_protected <= 0.1)/n()) %>% 
  left_join(ocean_matrix) %>% 
  select(x, y, fraction_in_top10) %>% 
  rasterFromXYZ(crs = crs(ocean_low_res_moll)) %>% 
  projectRaster(ocean_low_res_moll, method = 'ngb') %>% 
  mask(ocean_low_res_moll)

uncertainty_map_top10 <- tmap::tm_shape(uncertainty_raster_top10)+
  tmap::tm_raster(title = "",
                  palette  = viridis::viridis(n = 4, direction = 1),
                  breaks = c(0, 0.75, 0.9, 0.95, 1),
                  labels = c("<75", "75-90","90-95", ">95"),
                  legend.is.portrait = T, 
                  legend.reverse = T)+
  tmap::tm_shape(land_50_moll)+
  tmap::tm_fill(col = "grey", border.col = "transparent")+
  tmap::tm_shape(MPAs)+
  tmap::tm_credits("% of runs within top 10%")+
  tmap::tm_raster(col = "fully_highly_protected_reviewed_MPAs", palette = "lightblue", legend.show	= F)+
  tmap::tm_layout(title = "Biodiversity Conservation Priorities",
                  title.position = c("center", .95),
                  inner.margins = c(0.12, 0, 0.08, 0.04), 
                  frame = F,
                  legend.position = c(.99, "center"))

tmap::tmap_save(uncertainty_map_top10, 
                filename = file.path(variant_dir, "uncertainty_map_top10.png"),
                dpi = 300,  height = 4, width = 8 )
```
## Uncertainty in top 30%

```{r}
# Calculate fraction of runs in top 30%
og_top30_cells <- og_results %>% 
  filter(fraction_protected <= 0.3) %>% 
  pull(cell_id)

uncertainty_raster_top30 <- uncertainty_runs_results %>% 
  filter(cell_id %in% og_top30_cells) %>% 
  group_by(cell_id) %>% 
  summarise(fraction_in_top30 = sum(fraction_protected <= 0.3)/n()) %>% 
  left_join(ocean_matrix) %>% 
  select(x, y, fraction_in_top30) %>% 
  rasterFromXYZ(crs = crs(ocean_low_res_moll)) %>% 
  projectRaster(ocean_low_res_moll, method = 'ngb') %>% 
  mask(ocean_low_res_moll)

uncertainty_map_top30 <- tmap::tm_shape(uncertainty_raster_top30)+
  tmap::tm_raster(title = "",
                  palette  = viridis::viridis(n = 4, direction = 1),
                  breaks = c(0, 0.75, 0.9, 0.95, 1),
                  labels = c("<75", "75-90","90-95", ">95"),
                  legend.is.portrait = T, 
                  legend.reverse = T)+
  tmap::tm_shape(land_50_moll)+
  tmap::tm_fill(col = "grey", border.col = "transparent")+
  tmap::tm_shape(MPAs)+
  tmap::tm_credits("% of runs within top 30%")+
  tmap::tm_raster(col = "fully_highly_protected_reviewed_MPAs", palette = "lightblue", legend.show	= F)+
  tmap::tm_layout(title = "Biodiversity Conservation Priorities",
                  title.position = c("center", .95),
                  inner.margins = c(0.12, 0, 0.08, 0.04), 
                  frame = F,
                  legend.position = c(.99, "center"))

tmap::tmap_save(uncertainty_map_top30, 
                filename = file.path(variant_dir, "uncertainty_map_top30.png"),
                dpi = 300,  height = 4, width = 8 )
```

# Coefficient of variation

```{r}
cv_raster <- uncertainty_runs_results %>% 
  group_by(cell_id) %>% 
  summarise(cv = raster::cv(fraction_protected)) %>% 
  left_join(ocean_matrix) %>% 
  select(x, y, cv) %>% 
  rasterFromXYZ(crs = crs(ocean_low_res_moll)) %>% 
  projectRaster(ocean_low_res_moll, method = 'ngb') %>% 
  mask(ocean_low_res_moll)

cv_map <- tmap::tm_shape(cv_raster)+
  tmap::tm_raster(title = "CV",
                  palette  = viridis::inferno(n = 10),
                  legend.is.portrait = T, 
                  legend.reverse = T,
                  style = "cont")+
  tmap::tm_shape(land_50_moll)+
  tmap::tm_fill(col = "grey", border.col = "transparent")+
  tmap::tm_shape(MPAs)+
  tmap::tm_raster(col = "fully_highly_protected_reviewed_MPAs", palette = "lightblue", legend.show	= F)+
  tmap::tm_layout(title = "Uncertainty in Biodiversity Conservation Priorities",
                  title.position = c("center", .95),
                  inner.margins = c(0.12, 0, 0.08, 0.04), 
                  frame = F,
                  legend.position = c(.99, "center"))

tmap::tmap_save(cv_map, 
                filename = file.path(variant_dir, "cv_map.png"),
                dpi = 300,  height = 4, width = 8 )
```
